Update jquery.scrollTo.js
We really need this because when the `val` is `null` then condition returns `true` (because of `typeof null == 'object'`), and the `both` function returns `null` value which is set to `offset` variable. Later we have the following line #118 `attr[key] += offset[pos] || 0;` and as the `offset` is `null` we get error message (totally sad face goes here).